pyercot/pyproject.toml
[project]
name = "pyercot"
version = "0.1.0"
authors = [
    { name = "Kevin Kenyon", email = "kevin@poweredbylight.com" }
]
description = "A client library for accessing ERCOT Public API Client/Developer Documentation"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[tool.uv.build-backend]
module-name = "pyercot"
module-root = ""

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
